requirements-check.txt
# various code formatters
black==22.3.0
isort==5.10.1
docformatter==1.4
# flake8 + plugins
flake8==4.0.1
flake8-alfred==1.1.1
flake8-broken-line==0.4.0
flake8-bugbear==22.4.25
flake8-builtins==1.5.3
flake8-cognitive-complexity==0.1.0
flake8-comprehensions==3.8.0
flake8-eradicate==1.2.1
flake8-expression-complexity==0.0.11
flake8-functions==0.0.7
flake8-functions-names==0.1.0
flake8-printf-formatting==1.1.2
flake8-pytest-style==1.6.0
flake8-simplify==0.19.2
pep8-naming==0.12.1
# pre-commit
pre-commit>=2.20.0,<3.0
